VIPM: Determine required LV version from build specification
When building VI packages, the VI Package Manager will automatically
determine the right version for a package based on the information
configured in the specification. Optimally, the same version should be
used for other build jobs in the solution.

The version information is not exposed by the VIPM API, so it is read
directly from the package specification.