VIPM: Add icon support.
This uses the official icon that ships with VIPM for display in the
solution explorer. It does this by reading the icon file directly from
disk.

The current implementation only works for Windows OS and only when VIPM
is installed at the default location. Otherwise it defaults to no icon.