VIPM: Fix build and patch version numbers are swapped
The version numbers diplayed in Solution Explorer are not in the
right order.

Expected
  major.minor.patch.build

Actual
  major.minor.build.patch

This produces unexpected results when using version labels. Instead, the
version numbers should be set by index. Eventually, when the labels are
fixed, this will continue to work. This is also the same way other
libraries have implemented it.