Caraya: Add solution item and build step
This library implements build steps to run Caraya unit tests from
Solution Explorer. All source types supported by Caraya are also
supported by this library (path, project, lvlib, vi, etc...).

A JUnit test report is created in the temporary folder with the same
name as the file or folder under test. This file is not yet exposed
fully as an artifact for the pipeline.

The implementation is done by calling the source VI in <vi.lib>. This is
necessary so that Caraya can find unit tests in the source file or
directory, which is done by looking at the linker info. If the VIs were
build into the PPL, then it wouldn't be possible because it would
attempt to find unit tests for VIs in the PPL and not <vi.lib>. See the
build core method of the test step for more information about the (not
working) alternatives.