Caraya: Return test report as build artifact
This allows following steps to access the test report for further
processing.

This commit also adds the Caraya icon to Solution Explorer. This only
works on Windows and only when the installed package is still located
in the package cache of VIPM (this is a hack to prevent copying the
icon into the repository... for now).