solution: Add build step for Caraya
This creates two SFX archives, one for Caraya and VIPM, respectively. It
would be better to have a single installer but that is not possible due
to the dependencies (it is currently not possible to exclude items from
the SFX archive). Subfolders also don't work because Solution Explorer
only looks for plugins in the root folder (ignores sub-folders).