Include license files in all installers
This adds the entire licenses/ folder and the LICENSE.txt file to all
installers.

Note: Solution Explorer does not provide a way to filter the types of
files it will include from the licenses/ folder, so the .gitkeep file is
also included. Not pretty but still better than selecting each file
manually, which would be very error prone.